+++ src/main/java/com/packsendme/lib/simulation/airway/SimulationAirplaneResponse_Dto.java
@@ -1,4 +1,4 @@
-package com.packsendme.lib.simulation.way.maritimeway.response.dto;
+package com.packsendme.lib.simulation.airway;
 
 import java.io.Serializable;
 
@@ -9,11 +9,12 @@ import lombok.Setter;
 @Getter 
 @Setter 
 @NoArgsConstructor
-public class SimulationShipResponse_Dto implements Serializable  {
+public class SimulationAirplaneResponse_Dto implements Serializable  {
 	
 	/**
 	 * 
 	 */
 	public static final long serialVersionUID = 1L;
-	public Double cost_shipcontainer;
+	public Double cost_bag;	
+	public Boolean status_bag;
 }
