+++ src/main/java/com/packsendme/lib/simulation/maritimeway/SimulationMaritimewayResponse_Dto.java
@@ -1,8 +1,6 @@
-package com.packsendme.lib.simulation.way.response.dto;
+package com.packsendme.lib.simulation.maritimeway;
 
 import java.io.Serializable;
-
-import com.packsendme.lib.simulation.way.maritimeway.response.dto.SimulationShipResponse_Dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
