+++ src/main/java/com/packsendme/lib/simulation/maritimeway/SimulationShipResponse_Dto.java
@@ -1,4 +1,4 @@
-package com.packsendme.lib.simulation.way.airway.response.dto;
+package com.packsendme.lib.simulation.maritimeway;
 
 import java.io.Serializable;
 
@@ -9,12 +9,11 @@ import lombok.Setter;
 @Getter 
 @Setter 
 @NoArgsConstructor
-public class SimulationAirplaneResponse_Dto implements Serializable  {
+public class SimulationShipResponse_Dto implements Serializable  {
 	
 	/**
 	 * 
 	 */
 	public static final long serialVersionUID = 1L;
-	public Double cost_bag;	
-	public Boolean status_bag;
+	public Double cost_shipcontainer;
 }
